refactor: 결과를 매핑해주는 ResultSetExtractor 추가
