+++ jdbc/src/main/java/nextstep/jdbc/core/RowMapper.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.core;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
