+++ jdbc/src/main/java/nextstep/jdbc/core/RowMapperResultSetExtractor.java
@@ -1,0 +1,26 @@
+package nextstep.jdbc.core;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+import nextstep.jdbc.DataAccessException;
+
+public class RowMapperResultSetExtractor<T> implements ResultSetExtractor<List<T>> {
+
+    private final RowMapper<T> rowMapper;
+
+    public RowMapperResultSetExtractor(final RowMapper<T> rowMapper) {
+        this.rowMapper = rowMapper;
+    }
+
+    @Override
+    public List<T> extractData(final ResultSet rs) throws SQLException, DataAccessException {
+        int rowNum = 0;
+        List<T> results = new ArrayList<>();
+        while (rs.next()) {
+            results.add(rowMapper.mapRow(rs, rowNum++));
+        }
+        return results;
+    }
+}
